% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distDutch.R
\docType{data}
\name{distDutch}
\alias{distDutch}
\title{Results of applying VC-levenshtein distance on the \code{\link{Dutch}} dataset.}
\format{
A matrix with 613 rows and 613 variables:
\describe{
  \item{Aalsmeer NH}{A location in the list of data collection sites}
  ...
}
}
\source{
\url{https://gabmap.nl/examples/}
}
\usage{
distDutch
}
\description{
A dataset containing the results of applying VC-levenshtein distance on the \code{\link{Dutch}} dataset, which in turn is retrieved from the Goeman-Taeldeman-Van Reenen-Project.
}
\keyword{datasets}
