% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distmat_to_df.R
\name{distmat_to_df}
\alias{distmat_to_df}
\title{Convert a dialectometric distance matrix to a dataframe}
\usage{
distmat_to_df(dist_matrix)
}
\arguments{
\item{dist_matrix}{A distance matrix as generated by \code{\link{distance_matrix}}.}
}
\value{
A dataframe with the columns "row", "col", and "dist".
}
\description{
Input a distance matrix, with which a dataframe will be returned with the three columns of "row", "col", and "dist". The first two correspond with the rows and columns in the distance matrix, and the last refers to their crossing point, where the distance between them is given.
}
\examples{

## Example 1: Dutch distance matrix to Dutch dataframe
data(distDutch)
distmat_to_df(distDutch)
}
