% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dutch.R
\docType{data}
\name{Dutch}
\alias{Dutch}
\title{Results of applying VC-levenshtein distance on the \code{\link{Dutch}} dataset.}
\format{
A dataframe with 613 rows and 562 variables:
\describe{
  \item{aarde}{The first word in the word list used to compare dialects.}
  ...
}
}
\source{
\url{https://gabmap.nl/examples/}
}
\usage{
Dutch
}
\description{
A dataset containing the results of applying VC-levenshtein distance
}
\keyword{datasets}
