% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{DutchKML}
\alias{DutchKML}
\title{KML (Keyhole Markup Language) file for the \code{\link{Dutch}} dataset}
\source{
\url{https://gabmap.nl/examples/}
}
\description{
This is a a KML file which marks the locations where the \code{\link{Dutch}} dataset was collected.
}
\references{
\url{https://www.meertens.knaw.nl/projecten/mand/EGTRPdatata.html}
}
