% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-e.R
\docType{data}
\name{e}
\alias{e}
\title{Wav file of "e" in the international phonetic alphabet}
\format{
wav file.
}
\source{
\url{http://www.phonetics.ucla.edu/course/chapter1/vowels.html}
}
\description{
This is a wav file which contains the pronunciation of "e" in the international phonetic alphabet.
}
\examples{
i_audio <- system.file("extdata", "i.wav", package="dialectR")
e_audio <- system.file("extdata", "e.wav", package="dialectR")
try(acoustic_distance(i_audio, e_audio),
 message("Python not available for testing"),
 silent = TRUE)
}
