% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_points.R
\name{get_points}
\alias{get_points}
\title{Get KML points from KML data}
\usage{
get_points(kml_file_path)
}
\arguments{
\item{kml_file_path}{A file path to a KML file.}
}
\value{

}
\description{
Input a KML file path to get KML points data
}
