% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polygons.R
\name{get_polygons}
\alias{get_polygons}
\title{Get KML polygon from KML data}
\usage{
get_polygons(kml_file_path)
}
\arguments{
\item{kml_file_path}{A file path to a KML file.}
}
\value{

}
\description{
Input a KML file path to get KML polygon data
}
