% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{numpy}
\alias{numpy}
\alias{sklearn}
\alias{scipy}
\alias{python_speech_features}
\alias{speechpy}
\title{numpy module}
\format{
numpy module

sklearn module

scipy module

python_speech_features module

speechpy module
}
\usage{
numpy

sklearn

scipy

python_speech_features

speechpy
}
\description{
Interface to numpy module.

Interface to sklearn module.

Interface to scipy module.

Interface to python_speech_features module.

Interface to speechpy module.
}
\keyword{datasets}
