% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dutch.R
\docType{data}
\name{Dutch}
\alias{Dutch}
\title{Dutch dialect data from the Goeman-Taeldeman-Van Reenen-Project}
\format{
A dataframe with 613 rows and 562 variables
}
\source{
\url{https://gabmap.nl/examples/}
}
\usage{
Dutch
}
\description{
This dataset contains dialect data transcribed in the international phonetic alphabet, which is originally from the Goeman-Taeldeman-Van Reenen-Project.
}
\references{
\url{https://www.meertens.knaw.nl/projecten/mand/EGTRPdatata.html}
}
\keyword{datasets}
