% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-DutchKML.R
\docType{data}
\name{DutchKML}
\alias{DutchKML}
\title{KML (Keyhole Markup Language) file for the \code{\link{Dutch}} dataset}
\format{
KML file.
}
\source{
\url{https://gabmap.nl/examples/}
}
\description{
This is a KML file which marks the locations where the \code{\link{Dutch}} dataset was collected.
}
\examples{
dutch_points <- get_points(system.file("extdata", "DutchKML.kml", package="dialectR"))
dutch_points
}
\seealso{
\url{https://www.meertens.knaw.nl/projecten/mand/EGTRPdatata.html}
}
