% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{e}
\alias{e}
\title{Wav file of "e" in the international phonetic alphabet}
\source{
\url{http://www.phonetics.ucla.edu/course/chapter1/vowels.html}
}
\description{
This is a wav file which contains the pronunciation of "e" in the international phonetic alphabet, as made available by Peter Ladefoged for his textbook, A Course in Phonetics.
}
